Organize '/customers' page components in their respective folders
