+++ eisbuk-admin/src/pages/customers/CustomerSlots/CustomerSlots.stories.tsx
